Changed long integer type to support Windows compilation

+++ Block.h
@@ -14,17 +14,17 @@ namespace ADTFStream
 		{
 		  public:
 			// Member variables
-			unsigned long timeStamp = 0;
+			unsigned long long timeStamp = 0;
 			unsigned int fileIndexEntry = 0;
 			unsigned int previousOffset = 0;
 			unsigned int size = 0;
 			unsigned short streamId = 0;
 			unsigned short flags = 0;
-			unsigned long streamIndexEntry = 0;
+			unsigned long long streamIndexEntry = 0;
 
 			unsigned char sampleVersion = 0;
 			unsigned int sampleSize = 0;
-			unsigned long sampleTimeStamp = 0;
+			unsigned long long sampleTimeStamp = 0;
 			unsigned int sampleFlags = 0;
 		};
