+++ Extensions/Extension.h
@@ -18,7 +18,7 @@ namespace ADTFStream::Extensions
 		Extension(Header* header, FILE* file)
 			: header(header)
 		{
-			unsigned long offset = ftell(file);
+			unsigned long long offset = ftell(file);
 			fseek(file, header->dataOffset, SEEK_SET);
 			data = new T(file, header->dataSize);
 			fseek(file, offset, SEEK_SET);
