+++ Extensions/GUID.cpp
@@ -3,7 +3,7 @@
 
 using ADTFStream::Extensions::GUID;
 
-GUID::GUID(FILE* file, unsigned long dataSize)
+GUID::GUID(FILE* file, unsigned long long dataSize)
 {
 	guid = new char[dataSize];
 	fread(guid, dataSize, 1, file);
