+++ Extensions/IndexAdd.cpp
@@ -3,7 +3,7 @@
 
 using ADTFStream::Extensions::IndexAdd;
 
-IndexAdd::IndexAdd(FILE* file, unsigned long dataSize)
+IndexAdd::IndexAdd(FILE* file, unsigned long long dataSize)
 {
 	fread(this, sizeof(IndexAdd), 1, file);
 }
